chore(charts): reorganize Storybook into a clean Charts/* section
Replaces the scattered Lab/ChartV2* stories with a coherent Charts/ hierarchy
covering every mark and chrome component: Bar, Line, Area, Dot, Band,
Candlestick, Error Bar, Reference Line, WebGL (scatter/interactive/heatmap/
streaming), Color (auto-palette + accessor), Composite (mixed/kitchen-sink), and
Chrome (Axes & Grids, Legend, Swatch, Tooltip). Shared datasets live in _data.ts.
