+++ apps/storybook/stories/charts/AxesAndGrids.stories.tsx
@@ -2,22 +2,18 @@
 
 import type {Meta, StoryObj} from '@storybook/react';
 import {Chart, bar, ChartGrid, ChartAxis} from '@astryxdesign/charts';
+import {monthlyData} from './_data';
 
-interface AxesAndGridsArgs {
-  /** Show horizontal grid lines */
+interface Args {
   horizontalGrid: boolean;
-  /** Show vertical grid lines */
   verticalGrid: boolean;
-  /** Show the bottom (x) axis edge line */
   bottomAxisLine: boolean;
-  /** Show the left (y) axis edge line */
   leftAxisLine: boolean;
-  /** Show perpendicular tick marks on both axes */
   showTicks: boolean;
 }
 
-const meta: Meta<AxesAndGridsArgs> = {
-  title: 'Lab/Chart v2/Primitives/Axes & Grids',
+const meta: Meta<Args> = {
+  title: 'Charts/Chrome/Axes & Grids',
   argTypes: {
     horizontalGrid: {control: 'boolean'},
     verticalGrid: {control: 'boolean'},
@@ -35,26 +31,8 @@ const meta: Meta<AxesAndGridsArgs> = {
 };
 export default meta;
 
-const monthlyData = [
-  {month: 'Jan', revenue: 45},
-  {month: 'Feb', revenue: 52},
-  {month: 'Mar', revenue: 48},
-  {month: 'Apr', revenue: 61},
-  {month: 'May', revenue: 55},
-  {month: 'Jun', revenue: 70},
-];
-
-/**
- * Toggle grid lines, axis edge lines, and tick marks independently.
- *
- * - Grid lines use `--color-border-emphasized`; the y=0 / domain axis line
- *   uses `--color-icon-primary`.
- * - The bottom axis edge line is on by default; the left/top/right are off.
- * - When showing vertical grid lines, enable `leftAxisLine` so the grid
- *   has a clear left edge to anchor against.
- */
-export const AxesAndGrids: StoryObj<AxesAndGridsArgs> = {
-  name: 'Axes & Grids',
+/** Toggle grid lines, axis edge lines, and tick marks independently. */
+export const Playground: StoryObj<Args> = {
   render: args => (
     <Chart
       data={monthlyData}
